requirements.txt
streamlit
yfinance
requests
pandas 
PyPDF2
sentence_transformers 
numpy 
faiss-cpu